Start of generator rewrite
Attempt #1 at cleaning up the generators. A lot still needs to happen. There needs to be a more uniform way to execute the steps. Apparently dialog->getRunner is a thing.
I'm trying to hide all the input & output vars in an assistant object which is shared between the Commands and the Generators. The problem is that several generators are written completely differently from one another.